updated styling files to match Phaser template
