+++ client/index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Build-A-Terror</title>
     <!-- <link href="./src/output.css" rel="stylesheet"> -->
-    <link href="./src/gameUI.css" rel="stylesheet">
+    <link href="./src/style.css" rel="stylesheet">
  
 </head>
 <body class="">
