--- client/public/assets/style.css
@@ -1,6 +1,0 @@
-/* body {
-  margin: 0;
-  padding: 0;
-  color: rgba(255, 255, 255, 0.87);
-  background-color: #000000;
-} */
