+++ client/public/style.css
@@ -1,0 +1,281 @@
+/* Style for the header */
+
+* {
+  margin: 0;
+  padding: 0;
+}
+
+.game-header {
+  width: 100%;
+  height: 40px;
+  background-color: rgb(29, 38, 40);
+  color: rgb(213, 202, 163);
+  padding: 20px;
+  text-align: center;
+  font-size: 20px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  top: 0;
+  left: 0;
+  z-index: 10; /* Ensures it stays above other elements */
+}
+
+/* Make the title responsive */
+.game-header h1 {
+  font-size: 20px; /* Default font size for larger screens */
+  margin: 0;
+  padding: 0;
+  text-align: center; /* Align text to the left */
+}
+
+/* Media query for medium screens (tablets) */
+@media (max-width: 768px) {
+  .game-header {
+    flex-direction: column; /* Stack elements vertically */
+    align-items: flex-start;
+    padding: 10px;
+  }
+
+  .game-header h1 {
+    font-size: 20px; /* Adjust the font size for tablets */
+  }
+}
+
+/* Media query for small screens (phones) */
+@media (max-width: 480px) {
+  .game-header h1 {
+    font-size: 20px; /* Adjust the font size for phones */
+    text-align: left; /* Center the text for smaller screens */
+  }
+}
+
+h1 {
+  margin: 0; /* Remove default margin */
+  color: rgb(213, 202, 163);
+  font-size: 24px; /* Font size */
+}
+
+/* Style for buttons */
+.header-buttons {
+  display: flex;
+  gap: 15px; /* Space between buttons */
+  position: absolute;
+  right: 20px; /* Aligns buttons to the right */
+}
+
+.button {
+  background-color: rgb(213, 202, 163);
+  color: rgb(29, 38, 40);
+  border: none;
+  padding: 10px 20px;
+  font-size: 16px;
+  cursor: pointer;
+  border-radius: 5px; /* Rounded corners */
+  transition: background-color 0.3s; /* Smooth transition for hover effect */
+}
+
+.button:hover {
+  background-color: rgb(161, 153, 123);
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.3); /* Add a shadow on hover */
+}
+
+.button:focus {
+  outline: none; /* Remove focus outline */
+}
+
+/* Header layout for medium and small screens */
+@media (max-width: 768px) {
+  .header-buttons {
+    display: flex;
+    flex-direction: column; /* Stack buttons vertically */
+    gap: 10px; /* Space between buttons */
+    margin-top: 10px; /* Space between title and buttons */
+  }
+
+  .button {
+    padding: 8px 15px; /* Smaller padding on smaller screens */
+    font-size: 14px; /* Smaller font size */
+  }
+}
+
+/* For very small screens (e.g., phones in portrait mode) */
+@media (max-width: 480px) {
+  .button {
+    padding: 6px 12px; /* Even smaller padding */
+    font-size: 12px; /* Further reduce font size */
+  }
+}
+
+.container {
+  display: grid;
+  grid-template-columns: repeat(2, 1fr); /* Two equal columns */
+  width: 100%;
+}
+
+.left-container {
+  width: 50vw; /* Takes half of the viewport width */
+  height: 100vh; /* Full height */
+  background-image: url("/assets/greenBG.png"); /* Use the correct path */
+  background-size: cover; /* Ensures it covers the div */
+  background-position: center; /* Centers the image */
+  background-repeat: no-repeat; /* Prevents repeating */
+  text-align: center;
+  color: rgb(213, 202, 163);
+  position: relative; /* Needed for absolute positioning */
+  font-size: 20px; /* Default font size */
+}
+
+.left-container::before {
+  content: "";
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0, 0, 0, 0.4); /* Dark overlay */
+  z-index: 0; /* Places the overlay behind text */
+}
+
+.left-container h2,
+.left-container p {
+  position: relative; /* Ensures text appears above the overlay */
+  z-index: 1;
+}
+
+.right-container {
+  width: 50vw;
+  height: 100vh;
+  background-image: url("/assets/battlefieldBG.png");
+  background-size: cover;
+  background-position: center;
+  background-repeat: no-repeat;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  text-align: center;
+  color: rgb(29, 38, 40);
+  position: relative; /* Needed for absolute positioning */
+  font-size: 20px; /* Default font size */
+}
+
+/* Styling for battle-button */
+.battle-button {
+  background-color: rgb(213, 202, 163); /* Button color */
+  color: rgb(29, 38, 40); /* Text color */
+  border: none;
+  padding: 12px 24px;
+  font-size: 18px;
+  cursor: pointer;
+  border-radius: 8px; /* Rounded corners */
+  box-shadow: 2px 2px 5px rgba(0, 0, 0, 0.3);
+  transition: all 0.3s ease-in-out;
+  position: relative;
+  z-index: 10;
+}
+
+/* Hover effect */
+.battle-button:hover {
+  background-color: rgb(161, 153, 123); /* Darker shade */
+  box-shadow: 4px 4px 10px rgba(0, 0, 0, 0.5);
+}
+
+/* Button click (active) effect */
+.battle-button:active {
+  transform: scale(0.95);
+}
+
+/* Responsive button adjustment */
+@media (max-width: 768px) {
+  .right-container {
+    width: 100vw; /* Adjusts for mobile */
+    height: auto;
+    padding: 50px 20px;
+  }
+
+  .right-container .button {
+    width: 80%; /* Makes button more adaptive on smaller screens */
+  }
+}
+
+.right-container::before {
+  content: "";
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0, 0, 0, 0.4); /* Dark overlay */
+  z-index: 0; /* Places the overlay behind text */
+}
+
+.right-container h2,
+.right-container p {
+  position: relative; /* Ensures text appears above the overlay */
+  z-index: 1;
+}
+
+.game-footer {
+  width: 100%;
+  height: 40px;
+  background-color: rgb(29, 38, 40);
+  color: rgb(213, 202, 163);
+  padding: 20px;
+  text-align: center;
+  font-size: 15px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  top: 0;
+  left: 0;
+}
+
+/* Additional styling imported from Phaser Template */
+body {
+  margin: 0;
+  padding: 0;
+  color: rgba(255, 255, 255, 0.87);
+  background-color: #000000;
+  font-family: Arial, Helvetica, sans-serif;
+}
+
+#app {
+  width: 100%;
+  height: 100vh;
+  overflow: hidden;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.spritePosition {
+  margin: 10px 0 0 10px;
+  font-size: 0.8em;
+}
+
+.button {
+  width: 140px;
+  margin: 10px;
+  padding: 10px;
+  background-color: #000000;
+  color: rgba(255, 255, 255, 0.87);
+  border: 1px solid rgba(255, 255, 255, 0.87);
+  cursor: pointer;
+  transition: all 0.3s;
+
+  &:hover {
+    border: 1px solid #0ec3c9;
+    color: #0ec3c9;
+  }
+
+  &:active {
+    background-color: #0ec3c9;
+  }
+
+  /* Disabled styles */
+  &:disabled {
+    cursor: not-allowed;
+    border: 1px solid rgba(255, 255, 255, 0.3);
+    color: rgba(255, 255, 255, 0.3);
+  }
+}
